requirements.txt
# Core Dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.7.4,<3.0.0
pydantic-settings>=2.1.0

# Database
asyncpg>=0.29.0
psycopg2-binary>=2.9.0

# HTTP & API
httpx>=0.25.0
aiohttp>=3.9.0

# LLM & AI
openai>=1.10.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.3.3
langgraph>=0.4.5
langsmith>=0.2.10
mem0ai>=0.1.0

# Authentication & Security
pyjwt>=2.8.0

# Utilities
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.12.0
ruff>=0.1.0
mypy>=1.7.0